+++ components/admin/AdminPannel.tsx
@@ -7,7 +7,7 @@ import DashboardOverview from "@/components/admin/DashboardOverview"
 import ProfileSection from "@/components/admin/ProfileSection"
 import ProfileForm from "@/components/admin/Settings"
 import RoleManager from "@/components/auth/RoleManager"
-import EnrollStudent from "@/components/admin/AddStudent"
+import EnrollStudent from "@/components/admin/EnrollStudent"
 import NoDataAvailable from "./NoData"
 import IsLoading from "@/components/IsLoading"
 import { mobile, ipad } from "@/responsive"
